removed myaccount.py
